refactor(onboarding): move contextual options height into the CTA
Moves the options-content height clamp out of BrowserTabFragmentRenderer
and into BrandDesignContextualDaxDialogCta.applyContent(), so it runs on
every show path automatically (first-show, content transition, rotation
re-inflate) without needing the fragment to remember to call it.

Addresses https://github.com/duckduckgo/Android/pull/8439#discussion_r3225736286
