+++ app/src/main/java/com/duckduckgo/app/browser/BrowserTabFragment.kt
@@ -80,7 +80,6 @@ import androidx.core.view.ViewCompat
 import androidx.core.view.isGone
 import androidx.core.view.isVisible
 import androidx.core.view.postDelayed
-import androidx.core.view.updateLayoutParams
 import androidx.fragment.app.DialogFragment
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commitNow
@@ -5881,22 +5880,7 @@ class BrowserTabFragment :
                 existingRoot.addView(child)
             }
 
-            applyContextualOptionsHeight(cta)
             showCta(cta, instantShow = true)
-        }
-
-        private fun applyContextualOptionsHeight(cta: OnboardingDaxDialogCta.BrandDesignContextualDaxDialogCta) {
-            if (cta.activeIncludeId != R.id.contextualBrandDesignOptionsContent) return
-            val capHeight = resources.getBoolean(R.bool.capContextualOptionsHeight)
-            daxDialogInContextBrandDesign.root
-                .findViewById<View>(R.id.contextualBrandDesignOptionsContent)
-                ?.updateLayoutParams {
-                    height = if (capHeight) {
-                        resources.getDimensionPixelSize(R.dimen.contextualOptionsCappedHeight)
-                    } else {
-                        ViewGroup.LayoutParams.WRAP_CONTENT
-                    }
-                }
         }
 
         private fun showCta(configuration: Cta, instantShow: Boolean = false) {
@@ -5998,7 +5982,6 @@ class BrowserTabFragment :
                 }
                 // Both layouts live in the tree; only one should be visible at a time.
                 daxDialogInContext.root.gone()
-                applyContextualOptionsHeight(configuration)
                 configuration.showOnboardingCta(
                     binding,
                     { viewModel.onUserClickCtaOkButton(configuration) },
